requirements.txt
streamlit>=1.35
numpy>=1.24
scipy>=1.11
scikit-image>=0.21
matplotlib>=3.7
plotly>=5.18
pandas>=2.0
pydantic>=2.0
pyyaml>=6.0
pytest>=7.0
typing_extensions>=4.5
